backend/requirements.txt
# Python version: 3.10+
# Tested on: Python 3.10

fastapi~=0.104.0
uvicorn~=0.24.0
Pillow~=10.1.0
httpx~=0.25.0
openai~=1.6.0
python-dotenv~=1.0.0
aiosqlite~=0.19.0
zhdate~=0.1
dashscope~=1.14.0
